+++ src/main/java/com/clubber/ClubberServer/global/infrastructure/outer/api/oauth/kakao/client/KakaoInfoClient.java
@@ -1,4 +1,4 @@
-package com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.client.kakao;
+package com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.kakao.client;
 
 import com.clubber.ClubberServer.domain.auth.dto.UnlinkKaKaoTarget;
 import com.clubber.ClubberServer.global.config.feign.FeignConfig;
